requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
boto3==1.18.53
botocore==1.21.53
certifi==2021.5.30
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.6
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
daphne==3.0.2
decouple==0.0.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-environ==0.6.0
django-heroku==0.3.1
django-storages==1.11.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.17.1
gunicorn==20.1.0
hiredis==2.0.0
HttpPy==0.0.9
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
msgpack==1.0.2
oauthlib==3.1.1
Pillow==8.3.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
txaio==21.2.1
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zope.interface==5.4.0
